fix: refactor code issue: https://github.com/Telenav/osrm-backend/issues/143
