+++ integration/oasis/stationhandler/util.go
@@ -2,53 +2,12 @@ package stationhandler
 
 import (
 	"fmt"
-	"strconv"
 
 	"github.com/Telenav/osrm-backend/integration/oasis/osrmconnector"
+	"github.com/Telenav/osrm-backend/integration/oasis/osrmhelper"
 	"github.com/Telenav/osrm-backend/integration/pkg/api/osrm/coordinate"
-	"github.com/Telenav/osrm-backend/integration/pkg/api/osrm/table"
-	searchcoordinate "github.com/Telenav/osrm-backend/integration/pkg/api/search/coordinate"
-	"github.com/Telenav/osrm-backend/integration/pkg/api/search/nearbychargestation"
 	"github.com/golang/glog"
 )
-
-func generateSearchRequest(location searchcoordinate.Coordinate, limit int, radius float64) (*nearbychargestation.Request, error) {
-	// generate search request
-	req := nearbychargestation.NewRequest()
-	req.Location = location
-	if limit > 0 {
-		req.Limit = limit
-	}
-
-	if radius > 0 {
-		req.Radius = radius
-	}
-
-	return req, nil
-}
-
-func GenerateTableReq4Points(startPoints coordinate.Coordinates, endPoints coordinate.Coordinates) (*table.Request, error) {
-	if len(startPoints) == 0 || len(endPoints) == 0 {
-		return nil, fmt.Errorf("calling function with empty points")
-	}
-
-	// generate table request
-	req := table.NewRequest()
-	req.Coordinates = append(startPoints, endPoints...)
-
-	count := 0
-	for i := range startPoints {
-		str := strconv.Itoa(i)
-		req.Sources = append(req.Sources, str)
-		count++
-	}
-	for i := range endPoints {
-		str := strconv.Itoa(i + count)
-		req.Destinations = append(req.Destinations, str)
-	}
-
-	return req, nil
-}
 
 type cost struct {
 	duration float64
@@ -62,7 +21,7 @@ type costBetweenChargeStations struct {
 	err error
 }
 
-func calcCostBetweenChargeStationsPair2(from nearbyStationsIterator, to nearbyStationsIterator, oc *osrmconnector.OSRMConnector) <-chan costBetweenChargeStations {
+func calcCostBetweenChargeStationsPair(from nearbyStationsIterator, to nearbyStationsIterator, oc *osrmconnector.OSRMConnector) <-chan costBetweenChargeStations {
 	c := make(chan costBetweenChargeStations, 1000)
 
 	// collect (lat,lon)&ID for current location's nearby charge stations
@@ -90,7 +49,7 @@ func calcCostBetweenChargeStationsPair2(from nearbyStationsIterator, to nearbySt
 	}
 
 	// generate table request
-	req, err := GenerateTableReq4Points(startPoints, targetPoints)
+	req, err := osrmhelper.GenerateTableReq4Points(startPoints, targetPoints)
 	if err != nil {
 		closeChannelWithErrorInfo(c, err)
 		return c
